Updated code.

+++ src/main/java/com/santosh/product/dto/ResponseInfo.java
@@ -1,5 +1,7 @@
 package com.santosh.product.dto;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 import lombok.Data;
@@ -10,7 +12,7 @@ public class ResponseInfo {
 	
 	private long id;
 	private int productQuantity;
-	private double productPrice;
+	private BigDecimal productPrice;
 	private String productName;
