+++ src/main/java/com/santosh/product/entity/ProductEntity.java
@@ -1,13 +1,12 @@
 package com.santosh.product.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 
 import org.hibernate.annotations.ColumnDefault;
@@ -41,6 +40,6 @@ public class ProductEntity implements Serializable{
 	private int productQuantity;
 	
 	@Column(name = "PRDCT_PRICE")
-	private double productPrice;
+	private BigDecimal productPrice;
 
 }
